Fix DATA_FORMAT printf/scanf for 32-bit archs with 64-bit off_t off_t
On 32-bit architectures built with _FILE_OFFSET_BITS=64, off_t is 64-bit
but size_t is 32-bit. Using %zu for an off_t field causes scanf to write
only 4 bytes into an 8-byte off_t, corrupting the upper half.

Switch to %llu with explicit casts throughout.

+++ src/library/md5-backend.c
@@ -120,7 +120,8 @@ int add_file_to_backend_by_md5(const char *path, const char *expected_md5,
 	}
 
 	char *data;
-	if (asprintf(&data, DATA_FORMAT, trust_src, file_size, sha_digest) == -1) {
+	if (asprintf(&data, DATA_FORMAT, trust_src,
+		     (unsigned long long)file_size, sha_digest) == -1) {
 		data = NULL;
 	}
 	free(sha_digest);
@@ -153,4 +154,3 @@ int add_file_to_backend_by_md5(const char *path, const char *expected_md5,
 	}
 	return 1;
 }
-
